Adding Py3 support and other goodies
Python 3
=======
Python 3 support mainly required a change in how the data was passed
around - cherrypy is smart enough to distinguish between bytes and
chars (which is expected if it is providing Py3 support)

This also updated the page handler to conditionally return JSON back
dependent on content type. This logic *may* look brittle but it should
work just fine.

Other Stuff
========
- runnable unittest files
- requirements.txt!
- bug where function name was incorrect (stop_mock_service -> stop)
- PEP8 (especially 80char lines)
- Some refactoring where DRY-ness could be added
- simplejson was added, but not as a hard dependency